# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TDFF Backend — Grand Tour Cycling Fantasy League API",
    "version": "1.0.0",
    "description": "Real credential in production: a Bearer JWT issued by Auth0 (`AUTH_MODE=auth0`). Local dev only (`AUTH_MODE=dev`) accepts an unverified `x-user-id` header instead — not modeled here as a security scheme since it's not part of the real contract."
  },
  "servers": [
    {
      "url": "https://tdff-backend.vercel.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/competitions": {
      "post": {
        "operationId": "createCompetition",
        "tags": [
          "Competitions"
        ],
        "summary": "Create a competition",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Requires `admin` role. `slots` declares this competition instance’s required picks and their point values — e.g. GC Top 3 and KOM Top 3 can both use `top_1`/`top_2`/`top_3` at different point values; a stage-winner competition can be winner-only or top-3, per instance.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "type": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "fantasyLeagueId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "imageUrl": {
                    "type": "string",
                    "maxLength": 500,
                    "format": "uri"
                  },
                  "entryLockAt": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "date-time"
                  },
                  "slots": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "slot": {
                          "type": "string",
                          "enum": [
                            "overall_team",
                            "top_1",
                            "top_2",
                            "top_3",
                            "climber",
                            "sprinter",
                            "rouleur",
                            "puncheur"
                          ]
                        },
                        "points": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        }
                      },
                      "required": [
                        "slot",
                        "points"
                      ]
                    },
                    "minItems": 1
                  }
                },
                "required": [
                  "name",
                  "type",
                  "fantasyLeagueId",
                  "slots"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Competition created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "fantasyLeagueId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "entryLockAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "slots": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slot": {
                            "type": "string"
                          },
                          "points": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "slot",
                          "points"
                        ]
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "type",
                    "fantasyLeagueId",
                    "slots",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No valid credential on the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller doesn't have the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listCompetitions",
        "tags": [
          "Competitions"
        ],
        "summary": "List competitions (paginated)",
        "security": [],
        "description": "Defaults to 50 items/page (max 100).",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 100,
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of competitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "fantasyLeagueId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "imageUrl": {
                            "type": "string"
                          },
                          "entryLockAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "slots": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "slot": {
                                  "type": "string"
                                },
                                "points": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "slot",
                                "points"
                              ]
                            }
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "type",
                          "fantasyLeagueId",
                          "slots",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "limit",
                    "total",
                    "totalPages"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/competitions/{id}": {
      "get": {
        "operationId": "getCompetitionById",
        "tags": [
          "Competitions"
        ],
        "summary": "Get a competition by id",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Competition id"
            },
            "required": true,
            "description": "Competition id",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The competition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "fantasyLeagueId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "entryLockAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "slots": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slot": {
                            "type": "string"
                          },
                          "points": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "slot",
                          "points"
                        ]
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "type",
                    "fantasyLeagueId",
                    "slots",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateCompetitionDetails",
        "tags": [
          "Competitions"
        ],
        "summary": "Update a competition’s cosmetic details (currently just imageUrl)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Requires `admin` role. Partial update — omitted fields are left untouched. Unlike `PUT /:id/slots`, never blocked by an existing result: a banner image doesn’t affect scoring.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Competition id"
            },
            "required": true,
            "description": "Competition id",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageUrl": {
                    "type": "string",
                    "maxLength": 500,
                    "format": "uri"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated competition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "fantasyLeagueId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "entryLockAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "slots": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slot": {
                            "type": "string"
                          },
                          "points": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "slot",
                          "points"
                        ]
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "type",
                    "fantasyLeagueId",
                    "slots",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No valid credential on the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller doesn't have the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/competitions/{id}/slots": {
      "put": {
        "operationId": "updateCompetitionSlots",
        "tags": [
          "Competitions"
        ],
        "summary": "Reshape a competition’s required slots/points",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Requires `admin` role. Rejected with 409 once a result has been submitted for this competition — slots are locked in at that point.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Competition id"
            },
            "required": true,
            "description": "Competition id",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slots": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "slot": {
                          "type": "string",
                          "enum": [
                            "overall_team",
                            "top_1",
                            "top_2",
                            "top_3",
                            "climber",
                            "sprinter",
                            "rouleur",
                            "puncheur"
                          ]
                        },
                        "points": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        }
                      },
                      "required": [
                        "slot",
                        "points"
                      ]
                    },
                    "minItems": 1
                  }
                },
                "required": [
                  "slots"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated competition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "fantasyLeagueId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "entryLockAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "slots": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slot": {
                            "type": "string"
                          },
                          "points": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "slot",
                          "points"
                        ]
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "type",
                    "fantasyLeagueId",
                    "slots",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No valid credential on the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller doesn't have the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with existing state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}